% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGroupID.R
\name{addGroupID}
\alias{addGroupID}
\title{addGroupID}
\usage{
addGroupID(.data, name, ..., concor = F, all_combs = T)
}
\arguments{
\item{.data}{used in pipe setting}

\item{name}{non-character for the variable that will be added}

\item{...}{grouping variables}

\item{concor}{logical (defaults to FALSE). If using the function to create a separate
concordance dataset set to TRUE, otherwise the function adds the column to the existing data.}

\item{all_combs}{logical (defaults to TRUE). Only applicable when \code{concor = TRUE}. Setting to FALSE
only generates an ID for the combinations in the given data.}
}
\value{

}
\description{
addGroupID
}
\examples{
data(mtcars)

# add sequential identifier for
# `gear` called `id`
mtcars \%>\%
addGroupID(id, gear)
}
