% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropConstantCols.R
\name{dropConstantCols}
\alias{dropConstantCols}
\title{dropConstantCols}
\usage{
dropConstantCols(.data, verbose = F)
}
\arguments{
\item{.data}{input data as used in a pipe}

\item{verbose}{logical (defaults to FALSE). Setting to TRUE prints the dropped variables to the console.}
}
\value{
Version of \code{.data} where any columns that have no unique values are dropped.
}
\description{
dropConstantCols
}
\examples{
data(mtcars)
mtcars$constant <- "Same all along"
mtcars \%>\% dropConstantCols()
}
