% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Computes the greatest common divisor (GCD) of two integers (numbers),}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A number.}

\item{b}{A number.}
}
\value{
The largest number that divides \code{a} and \code{b}.
}
\description{
Computes the greatest common divisor (GCD) of two integers (numbers),
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)

}
\references{
\href{https://en.wikipedia.org/wiki/Euclidean_algorithm}{https://en.wikipedia.org/wiki/Euclidean_algorithm}
}
