% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BYM2.R
\name{getBYM2scale}
\alias{getBYM2scale}
\title{getBYM2scale}
\usage{
getBYM2scale(adj_mat)
}
\arguments{
\item{adj_mat}{fully connected binary weight matrix}
}
\value{
numeric scaling factor for BYM2 model
}
\description{
getBYM2scale
}
\examples{
W <- matrix(c(0,1,0,1,0,1,0,1,0), byrow = TRUE, ncol = 3)
getBYM2scale(W)
}
