% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jf.R
\name{jinvlogit}
\alias{jinvlogit}
\title{jinvlogit}
\usage{
jinvlogit(eta, upper = 1)
}
\arguments{
\item{eta}{vector of unconstrained variables to convert between 0 and upper}

\item{upper}{numeric scalar (defaults to 1). Useful when using a adjusted logit transformation}
}
\value{
vector of values constrained between 0 and \code{upper}
}
\description{
jinvlogit
}
\examples{
# Using standard implementation
x <- rnorm(10)
jlogit(x)

# Using values between 0 and 0.5
jinvlogit(x, upper = 0.5)
}
