% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jlogit.R
\name{jlogit}
\alias{jlogit}
\title{jlogit}
\usage{
jlogit(p, upper = 1)
}
\arguments{
\item{p}{vector of probabilities to convert to unconstrained space}

\item{upper}{numeric scalar (defaults to 1). Useful when using a adjusted logit transformation}
}
\value{
vector of unconstrained values
}
\description{
jlogit
}
\examples{
x <- runif(10)
jlogit(x)
}
