\name{MyFirstRPackage-package}
\alias{MyFirstRPackage-package}
\alias{MyFirstRPackage}
\docType{package}
\title{
\packageTitle{MyFirstRPackage}
}
\description{
\packageDescription{MyFirstRPackage}
}
\details{


\packageDESCRIPTION{MyFirstRPackage}
\packageIndices{MyFirstRPackage}

}
\author{
\packageAuthor{MyFirstRPackage}

\packageMaintainer{MyFirstRPackage}
}
\references{

}

\keyword{ package }
\seealso{
}
\examples{

}

