% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstRPackage.r
\docType{package}
\name{myFirstRPackage}
\alias{myFirstRPackage}
\title{Experimental package to learn how to create packages in R.}
\description{
myFirstRPackage package has two functions:
euclidean() and dijkstra().
}
\section{euclidean()}{

Computes the greatest common divisor (GCD) of two integers (numbers).
}

\section{dijkstra()}{

Finds the shortest paths between nodes in a graph
}

