% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/starPlot.R
\name{starPlot}
\alias{starPlot}
\title{Plot a Star}
\usage{
starPlot(col = "blue", add = T)
}
\arguments{
\item{col}{colour of the star}

\item{add}{Add to existing plot (T/F)}
}
\value{
Star plot
}
\description{
This function takes the data from datasauRus and
plot the star with you favorite color
}
