% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectorSumtoZero.R
\name{vectorSumtoZero}
\alias{vectorSumtoZero}
\title{vectorSumtoZero}
\usage{
vectorSumtoZero(n, fixed_sum = 0, mean = 0, sd = 1)
}
\arguments{
\item{n}{size of vector}

\item{fixed_sum}{defaults to zero}

\item{mean}{defaults to 0}

\item{sd}{defaults to 1}
}
\value{
numeric vector of length \code{n} from a normal distribution where
the sum of the vector is equal to \code{fixed_sum}. This is achieved by adjusting the final
observation.
}
\description{
vectorSumtoZero
}
\examples{
vectorSumtoZero(10)
}
