% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wiki_graph.r
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{wiki_graph}
\format{
A data frame with 18 rows and 3 variables:
\describe{
\item{v1}{first edge}
\item{v2}{second edge}
\item{w}{weight between edges}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Graph}
}
\usage{
wiki_graph
}
\description{
data.frame with three variables (v1, v2 and w) that contains the edges of the
graph (from v1 to v2) with the weight of the edge (w)
}
\keyword{datasets}
