% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUnique.R
\name{getUnique}
\alias{getUnique}
\title{getUnique}
\usage{
getUnique(.data)
}
\arguments{
\item{.data}{input data as used in a pipe}
}
\value{
Dataframe where the rows are the variables of \code{.data}. For each
variable the function returns the number of unique values, \code{n}, and what these
values are, \code{value}.
}
\description{
getUnique
}
\examples{
data(mtcars)
mtcars \%>\% getUnique()
}
