% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jf.R
\name{getUnique}
\alias{getUnique}
\title{getUnique}
\usage{
getUnique(.data)
}
\arguments{
\item{.data}{}
}
\value{
data frame with all combinations of the rows
}
\description{
getUnique
}
\examples{
x <- rnorm(10)
}
