% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myFirstRPackage-package.R
\docType{package}
\name{myFirstRPackage-package}
\alias{myFirstRPackage}
\alias{myFirstRPackage-package}
\title{myFirstRPackage: Everyday R solutions}
\description{
Several R functions that I use in my current R code.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://jamiehogg-depo.github.io/myFirstRPackage/}
}

}
\author{
\strong{Maintainer}: Jamie Hogg \email{james.hogg@hdr.qut.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
