\name{MyFirstRPackage-package}
\alias{MyFirstRPackage-package}
\alias{MyFirstRPackage}
\docType{package}
\title{
\packageTitle{MyFirstRPackage}
}
\description{
\packageDescription{MyFirstRPackage}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MyFirstRPackage}
\packageIndices{MyFirstRPackage}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MyFirstRPackage}

Maintainer: \packageMaintainer{MyFirstRPackage}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS ~~
~~ in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
#~~ simple examples of the most important functions ~~
}
